backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.3",
    "django-cors-headers>=4.7.0",
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
    "dotenv>=0.9.9",
    "pillow>=11.1.0",
    "requests>=2.32.5",
]
